+++ commerce/src/main/java/store/_0982/commerce/application/order/CanceledOrderService.java
@@ -18,6 +18,8 @@ import store._0982.commerce.domain.order.policy.ReversalOrderCancellationPolicy;
 import store._0982.commerce.domain.order.policy.VoidOrderCancellationPolicy;
 import store._0982.commerce.exception.CustomErrorCode;
 import store._0982.common.domain.grouppurchase.GroupPurchase;
+import store._0982.common.domain.order.CancelStatus;
+import store._0982.common.domain.order.CanceledOrder;
 import store._0982.common.domain.order.Order;
 import store._0982.common.domain.order.OrderStatus;
 import store._0982.common.exception.CustomException;
