+++ commerce/src/main/java/store/_0982/commerce/application/order/event/OrderCancelProcessedEvent.java
@@ -1,6 +1,6 @@
 package store._0982.commerce.application.order.event;
 
-import store._0982.commerce.domain.order.CanceledOrder;
+import store._0982.common.domain.order.CanceledOrder;
 
 public record OrderCancelProcessedEvent(
         CanceledOrder canceledOrder,
