+++ commerce/src/main/java/store/_0982/commerce/application/settlement/OrderSettlementService.java
@@ -3,7 +3,7 @@ package store._0982.commerce.application.settlement;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import store._0982.commerce.domain.order.CanceledOrder;
+import store._0982.common.domain.order.CanceledOrder;
 import store._0982.commerce.domain.settlement.OrderSettlementRepository;
 import store._0982.common.domain.order.Order;
 import store._0982.common.domain.settlement.OrderSettlement;
