+++ commerce/src/main/java/store/_0982/commerce/infrastructure/order/CanceledOrderJpaRepository.java
@@ -1,8 +1,8 @@
 package store._0982.commerce.infrastructure.order;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import store._0982.commerce.domain.order.CancelStatus;
-import store._0982.commerce.domain.order.CanceledOrder;
+import store._0982.common.domain.order.CancelStatus;
+import store._0982.common.domain.order.CanceledOrder;
 
 import java.time.OffsetDateTime;
 import java.util.Collection;
