+++ commerce/src/main/java/store/_0982/commerce/infrastructure/order/CanceledOrderRepositoryAdaptor.java
@@ -2,8 +2,8 @@ package store._0982.commerce.infrastructure.order;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import store._0982.commerce.domain.order.CancelStatus;
-import store._0982.commerce.domain.order.CanceledOrder;
+import store._0982.common.domain.order.CancelStatus;
+import store._0982.common.domain.order.CanceledOrder;
 import store._0982.commerce.domain.order.CanceledOrderRepository;
 
 import java.time.OffsetDateTime;
