+++ common/src/main/java/store/_0982/common/domain/order/CancelStatus.java
@@ -1,4 +1,4 @@
-package store._0982.commerce.domain.order;
+package store._0982.common.domain.order;
 
 public enum CancelStatus {
     REQUESTED,   // 취소 요청
